I forgor to ConfigManager::read_settings() 💀

+++ SourceSDK/CDOTACamera.h
@@ -3,16 +3,16 @@
 
 class CDOTACamera {
 public:
-    void set_distance(float distance) {
-        Memory::write_memory(this, distance);
+    void set_distance(int distance) {
+        Memory::write_memory(this, static_cast<float>(distance));
     }
 
-    void set_fow(float fow) {
-        Memory::write_memory(this + 0x4, fow);
+    void set_fow(int fow) {
+        Memory::write_memory(this + 0x4, static_cast<float>(fow));
     }
 
-    void set_r_farz(float r_farz) {
-        Memory::write_memory(this + 0x14, r_farz);
+    void set_r_farz(int r_farz) {
+        Memory::write_memory(this + 0x14, static_cast<float>(r_farz));
     }
 
     std::optional<float> get_distance() {
